refactor: extract table creation logic
